Moved all custom code to `src` folder
To keep things ordered all custom code that makes up our app got moved
into the `src` folder. By this I hope to get a good separation between
the rest of the app and not get things weirdly entangled.
