+++ index.js
@@ -1,3 +1,3 @@
 import { AppRegistry } from 'react-native';
-import App from './App';
+import App from './src/App';
 AppRegistry.registerComponent('kannji', () => App);
